requirements.txt
matplotlib>=2.1.0
pandas>=0.20.3
pyserial>=3.4
seaborn>=0.8.1
sh>=1.12.14
